:bug: Fix: Known bug with getBLock
