+++ configs/tsconfig/package.json
@@ -2,5 +2,7 @@
 	"name": "@tevm/tsconfig",
 	"version": "1.0.0-next.40",
 	"private": false,
-	"files": ["base.json"]
+	"files": [
+		"base.json"
+	]
 }
