+++ packages/blockchain/src/createBaseChain.js
@@ -1,9 +1,8 @@
 import { Block } from '@tevm/block'
-import { createJsonRpcFetcher } from '@tevm/jsonrpc'
 import { EMPTY_STATE_ROOT } from '@tevm/trie'
-import { numberToHex } from '@tevm/utils'
 import { hexToBytes } from 'viem'
 import { putBlock } from './actions/putBlock.js'
+import { getBlockFromRpc } from './utils/getBlockFromRpc.js'
 
 /**
  * Keccak-256 hash of the RLP of null
@@ -14,108 +13,6 @@ const KECCAK256_RLP_S = '0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622
  */
 const KECCAK256_RLP = hexToBytes(KECCAK256_RLP_S)
 
-// TODO move this to @tevm/block and make it coerse the type into type BlockTag if true
-/**
- * Determines if an unknown type is a valid block tag
- * @param {unknown} blockTag
- * @returns {boolean} true if valid block tag
- */
-const isBlockTag = (blockTag) => {
-	return typeof blockTag === 'string' && ['latest', 'earliest', 'pending', 'safe', 'finalized'].includes(blockTag)
-}
-
-// TODO move this to @tevm/block
-/**
- * @param {object} params
- * @param {string} params.url
- * @param {bigint | import('viem').BlockTag | import('viem').Hex} [params.blockTag]
- * @param {import('@tevm/common').Common} common
- */
-const getBlockFromRpc = async ({ url, blockTag = 'latest' }, common) => {
-	const fetcher = createJsonRpcFetcher(url)
-	/**
-	 * @param {import('viem').RpcBlock<'latest', true>} rpcBlock
-	 * @returns {Block}
-	 */
-	const asEthjsBlock = (rpcBlock) => {
-		return Block.fromRPC(
-			{
-				.../** @type {any}*/ (rpcBlock),
-				// filter out transactions we don't support as a hack
-				transactions: rpcBlock.transactions?.filter((tx) => {
-					// we currently don't support optimism deposit tx which uses this custom code
-					// Optimism type is currently not in viem types
-					// @ts-expect-error
-					if (tx.type === '0x7e') {
-						console.warn(
-							`Warning: Optimism deposit transactions (type 0x7e) are currently not supported and will be filtered out of blocks until support is added
-filtering out tx ${/** @type {import('viem').RpcBlock}*/ (tx).hash}`,
-						)
-						return false
-					}
-					return true
-				}),
-			},
-			undefined,
-			{ common, setHardfork: false, freeze: true, skipConsensusFormatValidation: true },
-		)
-	}
-	// TODO handle errors from fetch better
-	if (typeof blockTag === 'bigint') {
-		const { result, error } =
-			/** @type {{result: import('viem').RpcBlock<'latest', true>, error: {code: number | string, message: string}}}*/ (
-				await fetcher.request({
-					jsonrpc: '2.0',
-					id: 1,
-					method: 'eth_getBlockByNumber',
-					params: [numberToHex(blockTag), true],
-				})
-			)
-		if (error) {
-			throw error
-		}
-		if (!result) {
-			throw new Error('No block found')
-		}
-		return asEthjsBlock(result)
-	}
-	if (typeof blockTag === 'string' && blockTag.startsWith('0x')) {
-		const { result, error } = await fetcher.request({
-			jsonrpc: '2.0',
-			id: 1,
-			method: 'eth_getBlockByHash',
-			params: [blockTag, true],
-		})
-		if (error) {
-			console.error(error)
-			throw error
-		}
-		if (!result) {
-			console.error(error)
-			throw new Error('No block found')
-		}
-		return asEthjsBlock(/** @type {any}*/ (result))
-	}
-	if (isBlockTag(blockTag)) {
-		// TODO add an isBlockTag helper
-		const { result, error } = await fetcher.request({
-			jsonrpc: '2.0',
-			id: 1,
-			method: 'eth_getBlockByNumber',
-			params: [blockTag, true],
-		})
-		if (error) {
-			console.error(error)
-			throw error
-		}
-		if (!result) {
-			throw new Error('No block found')
-		}
-		return asEthjsBlock(/** @type {any}*/ (result))
-	}
-	throw new Error(`Invalid blocktag ${blockTag}`)
-}
-
 /**
  * Creates a genesis {@link Block} for the blockchain with params from {@link Common.genesis}
  * @param {Uint8Array} stateRoot The genesis stateRoot
@@ -123,52 +20,52 @@ filtering out tx ${/** @type {import('viem').RpcBlock}*/ (tx).hash}`,
  * @returns {Block}
  */
 const createGenesisBlock = (stateRoot, common) => {
-	const newCommon = common.copy()
-	newCommon.setHardforkBy({
-		blockNumber: 0,
-		td: BigInt(newCommon.genesis().difficulty),
-		timestamp: newCommon.genesis().timestamp ?? 0,
-	})
+  const newCommon = common.copy()
+  newCommon.setHardforkBy({
+    blockNumber: 0,
+    td: BigInt(newCommon.genesis().difficulty),
+    timestamp: newCommon.genesis().timestamp ?? 0,
+  })
 
-	/**
-	 * @type {import('@tevm/block').HeaderData}
-	 */
-	const header = {
-		...newCommon.genesis(),
-		number: 0,
-		stateRoot,
-		...(newCommon.isActivatedEIP(4895) ? { withdrawalsRoot: KECCAK256_RLP } : {}),
-	}
-	return Block.fromBlockData({ header, ...(newCommon.isActivatedEIP(4895) ? { withdrawals: [] } : {}) }, { common })
+  /**
+   * @type {import('@tevm/block').HeaderData}
+   */
+  const header = {
+    ...newCommon.genesis(),
+    number: 0,
+    stateRoot,
+    ...(newCommon.isActivatedEIP(4895) ? { withdrawalsRoot: KECCAK256_RLP } : {}),
+  }
+  return Block.fromBlockData({ header, ...(newCommon.isActivatedEIP(4895) ? { withdrawals: [] } : {}) }, { common })
 }
 /**
  * @param {import('./ChainOptions.js').ChainOptions} options
  * @returns {import('./BaseChain.js').BaseChain} Base chain object
  */
 export const createBaseChain = (options) => {
-	/**
-	 * @type {import('./BaseChain.js').BaseChain}
-	 */
-	const chain = {
-		options,
-		common: options.common,
-		blocks: new Map(),
-		blocksByTag: new Map(),
-		blocksByNumber: new Map(),
-		ready: () => genesisBlockPromise.then(() => true),
-	}
+  /**
+   * @type {import('./BaseChain.js').BaseChain}
+   */
+  const chain = {
+    options,
+    common: options.common,
+    blocks: new Map(),
+    blocksByTag: new Map(),
+    blocksByNumber: new Map(),
+    ready: () => genesisBlockPromise.then(() => true),
+  }
 
-	// Add genesis block and forked block to chain
-	const genesisBlockPromise = (async () => {
-		if (options.fork?.url) {
-			const block = await getBlockFromRpc(options.fork, options.common)
-			await putBlock(chain)(block)
-		} else {
-			await putBlock(chain)(
-				options.genesisBlock ?? createGenesisBlock(options.genesisStateRoot ?? EMPTY_STATE_ROOT, options.common),
-			)
-		}
-	})()
+  // Add genesis block and forked block to chain
+  const genesisBlockPromise = (async () => {
+    if (options.fork?.url) {
+      const block = await getBlockFromRpc(options.fork, options.common)
+      await putBlock(chain)(block)
+    } else {
+      await putBlock(chain)(
+        options.genesisBlock ?? createGenesisBlock(options.genesisStateRoot ?? EMPTY_STATE_ROOT, options.common),
+      )
+    }
+  })()
 
-	return chain
+  return chain
 }
